Further test refactoring
