+++ packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.toggleTwice.test.ts
@@ -2,30 +2,11 @@
  * @jest-environment jsdom
  */
 
-import { EditorSelection, EditorState, SelectionRange } from '@codemirror/state';
-import { EditorView } from '@codemirror/view';
+import { EditorSelection } from '@codemirror/state';
 import {
 	toggleCode, toggleMath, updateLink,
 } from './markdownCommands';
-import { GFM as GithubFlavoredMarkdownExt } from '@lezer/markdown';
-import { markdown } from '@codemirror/lang-markdown';
-import { MarkdownMathExtension } from './markdownMathParser';
-import { indentUnit } from '@codemirror/language';
-
-// Creates and returns a minimal editor with markdown extensions
-const createEditor = (initialText: string, initialSelection: SelectionRange): EditorView => {
-	return new EditorView({
-		doc: initialText,
-		selection: EditorSelection.create([initialSelection]),
-		extensions: [
-			markdown({
-				extensions: [MarkdownMathExtension, GithubFlavoredMarkdownExt],
-			}),
-			indentUnit.of('\t'),
-			EditorState.tabSize.of(4),
-		],
-	});
-};
+import createEditor from './createEditor';
 
 describe('markdownCommands, toggling twice', () => {
 	it('toggling math twice, starting on a line with content, should a math block', () => {
